requirements.txt
streamlit
pandas
sqlalchemy
mysql-connector-python
python-dotenv
scikit-learn
plotly